2.21.0
This is the final commit for the 2.21.0
In this version, The Nytuo Launcher will going to use the semantic versioning (MAJOR.MINOR.PATCH) like before.
This update bring new change to the update system, all is now automated on Github with Actions. The build, release and dependencies updates are automated.
This version broke the update flow for old Nytuo-Launcher versions.